ethiolostfound/items/serializers.py
from rest_framework import serializers
from .models import Item, CustomUser

class ItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = Item
        fields = '__all__'

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = [ 'name', 'photo']
