ethiolostfound/users/urls.py
from django.urls import path
from .views import RegisterView, CustomTokenObtainPairView
from rest_framework_simplejwt.views import TokenRefreshView,TokenVerifyView
from rest_framework_simplejwt.views import TokenBlacklistView

urlpatterns = [
    path('signup/', RegisterView.as_view(), name='signup'),  
    path('login/', CustomTokenObtainPairView.as_view(), name='login'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/token/verify/', TokenVerifyView.as_view(), name='token_verify'),
    path('api/token/blacklist/', TokenBlacklistView.as_view(), name='token_blacklist'),
]
